implement hitting and view events commands, add game engine responses

+++ src/core/game_engine.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import List, Optional
 
-from src.core.displayable_error import DisplayableError
+from src.core.game_engine_response import GameEngineResponse
 from src.entities.combo import Combo
 from src.entities.combo_set import ComboSet
 from src.entities.entry import Entry
@@ -19,30 +19,39 @@ class GameEngine:
         self.game_repo = game_repo
         self.player_repo = player_repo
 
-    def start_game(self, server_id: str, event_strs: List[str]) -> None:
+    def start_game(self, server_id: str, event_strs: List[str]) \
+            -> GameEngineResponse:
         game = self.game_repo.read_active_game(server_id)
         if game is not None:
-            raise DisplayableError('A game is already running.')
+            return GameEngineResponse(
+                display_error='A game is already running.')
 
         game = self.game_repo.create_game(server_id)
-        self.event_repo.create_events(game.game_id, event_strs)
+        events = self.event_repo.create_events(game.game_id, event_strs)
+        return GameEngineResponse({'events': events})
 
-    def finish_game(self, server_id: str) -> None:
+    def finish_game(self, server_id: str) -> GameEngineResponse:
         game = self.game_repo.read_active_game(server_id)
         if game is None:
-            raise DisplayableError('No game is currently running.')
+            return GameEngineResponse(
+                display_error='No game is currently running.')
 
         game.time_finished = datetime.now()
         self.game_repo.update_game(game)
+        return GameEngineResponse()
 
     def set_entry(self, server_id: str, player_id: str,
-                  combo_set_indices: List[List[int]]) -> None:
+                  combo_set_indices: List[List[int]]) -> GameEngineResponse:
         game = self.game_repo.read_active_game(server_id)
+        if game is None:
+            return GameEngineResponse(
+                display_error='No game is currently running.')
+
         events = self.event_repo.read_all_events(game.game_id)
         for event in events:
             if event.is_hit:
-                raise DisplayableError(
-                    'Cannot add/change entry as events have already been hit')
+                return GameEngineResponse(
+                    display_error='Cannot add/change entry as events have already been hit')
 
         self.player_repo.delete_entry(game.game_id, player_id)
 
@@ -52,10 +61,11 @@ class GameEngine:
                 combos.append(Combo([events[i]
                               for i in event_indices], combo_index))
             except IndexError:
-                raise DisplayableError('Invalid event index.')
+                return GameEngineResponse(display_error='Invalid event index.')
         combo_set = ComboSet(player_id, combos)
 
         self.player_repo.create_entry(game.game_id, combo_set)
+        return GameEngineResponse()
 
     def _search_event(self, game_id, *, index: Optional[int],
                       desc: Optional[str]) -> Event:
@@ -63,13 +73,12 @@ class GameEngine:
             hit_event = self.event_repo.read_event_by_index(
                 game_id, index)
             if hit_event is None:
-                raise DisplayableError('Invalid event index.')
+                return GameEngineResponse(display_error='Invalid event index.')
         elif desc is not None:
             hit_events = self.event_repo.read_events_by_desc(
                 game_id, desc)
             if len(hit_events) == 0:
-                raise DisplayableError(
-                    'No event description matches search string.')
+                return GameEngineResponse(display_error='No event description matches search string.')
             elif len(hit_events) > 1:
                 error_str = (
                     'More than one event description matches search string.\n'
@@ -79,30 +88,46 @@ class GameEngine:
                 )
                 if len(hit_events) > 2:
                     error_str += '\n...'
-                raise DisplayableError(error_str)
+                return GameEngineResponse(display_error=error_str)
             else:
                 hit_event = hit_events[0]
         else:
             raise ValueError('No event has been specified.')
-        return hit_event
+        return GameEngineResponse({'event': hit_event})
 
     def hit(self, server_id: str, *, index: int = None, desc: str = None) \
             -> None:
         game = self.game_repo.read_active_game(server_id)
-        hit_event = self._search_event(game.game_id, index=index, desc=desc)
+        if game is None:
+            return GameEngineResponse(
+                display_error='No game is currently running.')
+
+        response = self._search_event(game.game_id, index=index, desc=desc)
+        if response.display_error is not None:
+            return response
+        hit_event = response.response['event']
         if hit_event.is_hit:
-            raise DisplayableError((f'Event "{hit_event.desc}" ',
-                                    'has already been hit.'))
+            return GameEngineResponse(
+                display_error=f'Event "{hit_event.desc}" has already been hit.'
+            )
         hit_event.is_hit = True
         self.event_repo.update_event(game.game_id, hit_event)
+        return GameEngineResponse({'event': hit_event})
 
     def unhit(self, server_id: str, *, index: int = None, desc: str = None) \
             -> None:
         game = self.game_repo.read_active_game(server_id)
-        unhit_event = self._search_event(game.game_id, index=index, desc=desc)
+        if game is None:
+            return GameEngineResponse(
+                display_error='No game is currently running.')
+
+        response = self._search_event(game.game_id, index=index, desc=desc)
+        if response.display_error is not None:
+            return response
+        unhit_event = response.response['event']
         if not unhit_event.is_hit:
-            raise DisplayableError((f'Event "{unhit_event.desc}" ',
-                                    'is already unhit.'))
+            return GameEngineResponse(
+                display_error=f'Event "{unhit_event.desc}" is already unhit.')
         unhit_event.is_hit = False
         self.event_repo.update_event(game.game_id, unhit_event)
 
@@ -115,22 +140,34 @@ class GameEngine:
             if entry.time_won is not None and not combo_set.has_won():
                 entry.time_won = None
                 self.player_repo.update_entry(entry)
+        return GameEngineResponse({'event': unhit_event})
 
     def bingo(self, server_id: str, player_id: str) -> None:
         game = self.game_repo.read_active_game(server_id)
+        if game is None:
+            return GameEngineResponse(
+                display_error='No game is currently running.')
+
         entry = self.player_repo.read_entry(game.game_id, player_id)
         if entry.time_won is not None:
-            raise DisplayableError('Entry has already won.')
+            return GameEngineResponse(display_error='Entry has already won.')
 
         combo_set = self.player_repo.read_combo_set(game.game_id, player_id)
         if combo_set.has_won():
             entry.time_won = datetime.now()
             self.player_repo.update_entry(entry)
         else:
-            raise DisplayableError('Entry has not yet won.')
+            return GameEngineResponse(display_error='Entry has not yet won.')
+        return GameEngineResponse()
 
     def view_events(self, server_id: str) -> None:
-        pass
+        game = self.game_repo.read_active_game(server_id)
+        if game is None:
+            return GameEngineResponse(
+                display_error='No game is currently running.')
+
+        events = self.event_repo.read_all_events(game.game_id)
+        return GameEngineResponse({'events': events})
 
     def view_progress(self, server_id: str) -> None:
         pass
